fix(common): move ReverseContractUpdate to `state_update` module
